requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
geojson==2.5.0
geomet==0.3.0
html2text==2020.1.16
idna==2.10
importlib-metadata==4.4.0
pkg-resources==0.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
requests==2.25.1
sentinelsat==1.0.1
six==1.16.0
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
